foreign_remittance_cal/pyproject.toml
[tool.poetry]
name = "hdfc"
version = "0.1.0"
description = ""
authors = ["silverscorpio <rajat.newtoneinstein@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
